Change package name in auth-service
