+++ auth-service/src/main/java/by/modsen/authservice/KeycloakServiceApplication.java
@@ -1,0 +1,13 @@
+package by.modsen.authservice;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class KeycloakServiceApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(KeycloakServiceApplication.class, args);
+    }
+
+}
