+++ auth-service/src/main/java/by/modsen/authservice/dto/ErrorDto.java
@@ -1,0 +1,6 @@
+package by.modsen.authservice.dto;
+
+public record ErrorDto(
+        int statusCode,
+        String message
+) {}
