+++ auth-service/src/main/java/by/modsen/authservice/dto/LoginRequest.java
@@ -1,0 +1,6 @@
+package by.modsen.authservice.dto;
+
+public record LoginRequest(
+        String username,
+        String password
+) {}
