+++ auth-service/src/main/java/by/modsen/authservice/dto/NewUserDto.java
@@ -1,0 +1,9 @@
+package by.modsen.authservice.dto;
+
+public record NewUserDto(
+        String username,
+        String email,
+        String password,
+        String firstName,
+        String lastName
+) {}
