+++ auth-service/src/main/java/by/modsen/authservice/dto/UserRole.java
@@ -1,0 +1,6 @@
+package by.modsen.authservice.dto;
+
+public enum UserRole {
+    DRIVER,
+    PASSENGER
+}
