+++ auth-service/src/main/java/by/modsen/authservice/exception/UserCreateException.java
@@ -1,0 +1,7 @@
+package by.modsen.authservice.exception;
+
+public class UserCreateException extends RuntimeException {
+    public UserCreateException(String message) {
+        super(message);
+    }
+}
