+++ auth-service/src/main/java/by/modsen/authservice/service/GroupService.java
@@ -1,0 +1,8 @@
+package by.modsen.authservice.service;
+
+public interface GroupService {
+
+    void assignGroup(String userId, String groupId);
+
+    void deleteGroupFromUser(String userId, String groupId);
+}
