+++ auth-service/src/main/java/by/modsen/authservice/service/RoleService.java
@@ -1,0 +1,8 @@
+package by.modsen.authservice.service;
+
+public interface RoleService {
+
+    void assignRole(String userId, String roleName);
+
+    void deleteFromUser(String userId, String roleName);
+}
