+++ auth-service/src/main/java/by/modsen/authservice/service/UserRegistrationService.java
@@ -1,0 +1,13 @@
+package by.modsen.authservice.service;
+
+import by.modsen.authservice.dto.LoginRequest;
+import by.modsen.authservice.dto.NewUserDto;
+import by.modsen.authservice.dto.UserRole;
+import org.keycloak.representations.AccessTokenResponse;
+
+public interface UserRegistrationService {
+
+    AccessTokenResponse getJwt(LoginRequest loginRequest);
+
+    void registerUser(NewUserDto newUserDto, UserRole userRole);
+}
