+++ auth-service/src/main/java/by/modsen/authservice/service/UserService.java
@@ -1,0 +1,21 @@
+package by.modsen.authservice.service;
+
+import by.modsen.authservice.dto.NewUserDto;
+import org.keycloak.admin.client.resource.UserResource;
+import org.keycloak.representations.idm.GroupRepresentation;
+import org.keycloak.representations.idm.RoleRepresentation;
+
+import java.util.List;
+
+public interface UserService {
+
+    String createUser(NewUserDto newUserDto);
+
+    void deleteUser(String userId);
+
+    UserResource getUser(String userId);
+
+    List<RoleRepresentation> getUserRoles(String userId);
+
+    List<GroupRepresentation> getUserGroups(String userId);
+}
