--- auth-service/src/main/java/by/modsen/keycloakservice/KeycloakServiceApplication.java
@@ -1,13 +1,0 @@
-package by.modsen.keycloakservice;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class KeycloakServiceApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(KeycloakServiceApplication.class, args);
-    }
-
-}
