--- auth-service/src/main/java/by/modsen/keycloakservice/config/WebSecurityConfig.java
@@ -1,37 +1,0 @@
-package by.modsen.keycloakservice.config;
-
-import by.modsen.keycloakservice.dto.UserRole;
-import lombok.RequiredArgsConstructor;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.security.config.annotation.method.configuration.EnableMethodSecurity;
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-import org.springframework.security.config.annotation.web.configurers.AbstractHttpConfigurer;
-import org.springframework.security.web.SecurityFilterChain;
-
-@Configuration
-@EnableWebSecurity
-@EnableMethodSecurity
-@RequiredArgsConstructor
-public class WebSecurityConfig {
-
-    private final JwtAuthConverter jwtAuthConverter;
-
-    @Bean
-    public SecurityFilterChain securityFilterChain(HttpSecurity http) throws Exception {
-        http
-                .authorizeHttpRequests(auth -> auth
-                        .requestMatchers("/api/v1/auth/**").permitAll()
-                        .requestMatchers("/test/passenger").hasRole(UserRole.PASSENGER.name())
-                        .requestMatchers("/test/driver").hasRole(UserRole.DRIVER.name())
-                        .anyRequest().authenticated()
-                )
-                .oauth2ResourceServer(oauth2 -> oauth2
-                        .jwt(jwt -> jwt.jwtAuthenticationConverter(jwtAuthConverter))
-                )
-                .csrf(AbstractHttpConfigurer::disable);
-
-        return http.build();
-    }
-}
