--- auth-service/src/main/java/by/modsen/keycloakservice/controller/UserController.java
@@ -1,44 +1,0 @@
-package by.modsen.keycloakservice.controller;
-
-import by.modsen.keycloakservice.dto.NewUserDto;
-import by.modsen.keycloakservice.service.UserService;
-import lombok.RequiredArgsConstructor;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-@RequestMapping("/users")
-@RequiredArgsConstructor
-public class UserController {
-
-    private final UserService userService;
-
-    @PostMapping
-    public ResponseEntity<?> createUser(@RequestBody NewUserDto newUserDto) {
-        userService.createUser(newUserDto);
-        return ResponseEntity.status(HttpStatus.CREATED).build();
-    }
-
-    @DeleteMapping("/{userId}")
-    public ResponseEntity<?> deleteUser(@PathVariable String userId) {
-        userService.deleteUser(userId);
-        return ResponseEntity.noContent().build();
-    }
-
-    @GetMapping("/{id}/roles")
-    public ResponseEntity<?> getUserRoles(@PathVariable String id) {
-        return ResponseEntity.status(HttpStatus.OK).body(userService.getUserRoles(id));
-    }
-
-    @GetMapping("/{id}/groups")
-    public ResponseEntity<?> getUserGroups(@PathVariable String id) {
-        return ResponseEntity.status(HttpStatus.OK).body(userService.getUserGroups(id));
-    }
-}
