--- auth-service/src/main/java/by/modsen/keycloakservice/dto/ErrorDto.java
@@ -1,6 +1,0 @@
-package by.modsen.keycloakservice.dto;
-
-public record ErrorDto(
-        int statusCode,
-        String message
-) {}
