--- auth-service/src/main/java/by/modsen/keycloakservice/dto/LoginRequest.java
@@ -1,6 +1,0 @@
-package by.modsen.keycloakservice.dto;
-
-public record LoginRequest(
-        String username,
-        String password
-) {}
