--- auth-service/src/main/java/by/modsen/keycloakservice/dto/NewUserDto.java
@@ -1,9 +1,0 @@
-package by.modsen.keycloakservice.dto;
-
-public record NewUserDto(
-        String username,
-        String email,
-        String password,
-        String firstName,
-        String lastName
-) {}
