--- auth-service/src/main/java/by/modsen/keycloakservice/dto/UserRole.java
@@ -1,6 +1,0 @@
-package by.modsen.keycloakservice.dto;
-
-public enum UserRole {
-    DRIVER,
-    PASSENGER
-}
