--- auth-service/src/main/java/by/modsen/keycloakservice/exception/UserCreateException.java
@@ -1,7 +1,0 @@
-package by.modsen.keycloakservice.exception;
-
-public class UserCreateException extends RuntimeException {
-    public UserCreateException(String message) {
-        super(message);
-    }
-}
