--- auth-service/src/main/java/by/modsen/keycloakservice/service/GroupService.java
@@ -1,8 +1,0 @@
-package by.modsen.keycloakservice.service;
-
-public interface GroupService {
-
-    void assignGroup(String userId, String groupId);
-
-    void deleteGroupFromUser(String userId, String groupId);
-}
