--- auth-service/src/main/java/by/modsen/keycloakservice/service/RoleService.java
@@ -1,8 +1,0 @@
-package by.modsen.keycloakservice.service;
-
-public interface RoleService {
-
-    void assignRole(String userId, String roleName);
-
-    void deleteFromUser(String userId, String roleName);
-}
