+++ auth-service/src/test/java/by/modsen/authservice/UserServiceApplicationTests.java
@@ -1,0 +1,13 @@
+package by.modsen.authservice;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class UserServiceApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
