--- auth-service/src/test/java/by/modsen/keycloakservice/UserServiceApplicationTests.java
@@ -1,13 +1,0 @@
-package by.modsen.keycloakservice;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class UserServiceApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
